Extract automation_settings helpers into shared/settings.ts
getSetting/getRequiredSetting/getSettings read the automation_settings
collection and have nothing to do with email — they only lived in
email-service.ts by accident. Move them to a dedicated shared/settings.ts
with a minimal SettingsContext type (any hook context or EmailServiceContext
satisfies it structurally) and update all five importers. No behavior change.
