+++ directus-cms/extensions/directus-extension-programmierbar-bundle/src/shared/email-service.ts
@@ -5,7 +5,8 @@
  * - Template rendering using Handlebars syntax
  * - Email sending via Directus MailService
  * - Reading templates from the email_templates collection
- * - Reading settings from the automation_settings collection
+ *
+ * Reading values from the automation_settings collection lives in `settings.ts`.
  */
 
 import type { Logger } from 'pino'
@@ -69,89 +70,6 @@ export interface SendEmailOptions {
     replyTo?: string
     cc?: string
     attachments?: EmailAttachment[]
-}
-
-/**
- * Get a setting value from the automation_settings collection.
- */
-export async function getSetting(
-    key: string,
-    context: EmailServiceContext
-): Promise<string | null> {
-    const { services, getSchema, logger } = context
-    const { ItemsService } = services
-
-    try {
-        const schema = await getSchema()
-        const settingsService = new ItemsService('automation_settings', {
-            schema,
-            accountability: { admin: true },
-        })
-
-        const settings = await settingsService.readByQuery({
-            filter: { key: { _eq: key } },
-            fields: ['value'],
-            limit: 1,
-        })
-
-        if (settings && settings.length > 0) {
-            return settings[0].value
-        }
-
-        return null
-    } catch (err: any) {
-        logger.warn(`Failed to read setting '${key}': ${err?.message || err}`)
-        return null
-    }
-}
-
-/**
- * Like {@link getSetting}, but throws when the setting is missing or blank.
- *
- * Use this for configuration the extension cannot sensibly default (per
- * AGENTS.md: no fallback values for critical config — fail explicitly rather
- * than silently substituting a value that may be wrong for the environment).
- */
-export async function getRequiredSetting(key: string, context: EmailServiceContext): Promise<string> {
-    const value = await getSetting(key, context)
-    if (!value || !value.trim()) {
-        throw new Error(`Required setting '${key}' is not configured`)
-    }
-    return value
-}
-
-/**
- * Get multiple settings as an object.
- */
-export async function getSettings(
-    keys: string[],
-    context: EmailServiceContext
-): Promise<Record<string, string>> {
-    const { services, getSchema, logger } = context
-    const { ItemsService } = services
-
-    const result: Record<string, string> = {}
-
-    try {
-        const schema = await getSchema()
-        const settingsService = new ItemsService('automation_settings', {
-            schema,
-            accountability: { admin: true },
-        })
-
-        const settings = await settingsService.readByQuery({
-            filter: { key: { _in: keys } },
-            fields: ['key', 'value'],
-        })
-
-        for (const setting of settings) {
-            result[setting.key] = setting.value
-        }
-    } catch (err: any) {
-        logger.warn(`Failed to read settings: ${err?.message || err}`)
-    }
-
-    return result
 }
 
 /**
