+++ directus-cms/extensions/directus-extension-programmierbar-bundle/src/ticket-order-processing/index.ts
@@ -1,7 +1,8 @@
 import { defineHook } from '@directus/extensions-sdk'
 import { randomUUID } from 'node:crypto'
 import { Readable } from 'node:stream'
-import { sendTemplatedEmail, getSetting, type EmailServiceContext } from '../shared/email-service.js'
+import { sendTemplatedEmail, type EmailServiceContext } from '../shared/email-service.js'
+import { getSetting } from '../shared/settings.js'
 import { generateUniqueTicketCode, formatPrice } from '../shared/ticket-utils.js'
 import {
     generateInvoicePdf,
